Rename data_loader to data_fetcher
To clarify that this is about fetching external datasets and avoid
confusion with the loader module.
